+++ src/lib.rs
@@ -1,4 +1,5 @@
 #![no_std] // don't link the Rust standard library
+#![feature(abi_x86_interrupt)]
 
 extern crate bootloader_precompiled;
 extern crate spin;
@@ -13,9 +14,11 @@ extern crate array_init;
 #[cfg(test)]
 extern crate std;
 
-pub mod gdt;
-pub mod serial;
+#[macro_use]
 pub mod vga_buffer;
+pub mod gdt;
+pub mod interrupts;
+pub mod serial;
 
 pub unsafe fn exit_qemu() {
     use x86_64::instructions::port::Port;
